requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.15.0
requests>=2.31.0
base64
pathlib
typing
dataclasses
enum34
sqlite3
json
io
hashlib
datetime

